enh(ui): prevent SVideo crash on empty image update
When providing an empty image (size (0,0,0) ) to a SVideo service, the service does not filter and forwards it to the textureManager which tests it and generates an error in particular in debug mode (ASSERT).

However empty images are often used as default ones, and this leads to various difficulties.

Here, an if(size\[0\] == 0 || size\[1\] == 0) test is added on the SVideo updating and return when it fails.

Closes #1038