enh(viz): add point selection feature on SLandmarks
The SLandmark service is a graphical widget which handles the Landmarks, which can be seen as a vector of points. It has a feature to send the coordinates of the current selected point through a signal. 

This mechanism is enhanced, to treat a newly created point as a selected one. 
In addition, the selected point are outputed in a ```sight::data::Point``` variable in addition to  the signal, to make it possible to treat it as a classical sight data. 

closes 1040