enh(ui): add swapp support from single volume view to 3/4 view with negatos
The classical volume view in sightviewer is a single 3D scene with the volume. It is possible to go to a view with 4 scenes, with the 3D volume view on the top left, and the negato in all 3 directions in the other corner. 
It is presently not possible to simply make one of the negato in full screen.  

To make this possible, a SEvent service is added, to catch the different events, and forward a particular one. 
Once added to the configuration it is used to catch double click, and hide/show the different scenes. 
This makes it a lot easier to change the view, and go back. 

To keep the classical button functionality, the IAction api has been enhanced to support toogleVisibility. 

Closes #566