enh(viz): add default tf loading option in TransferFunctionWidget
The transferFunction editor widget (and windows) don't zllow to freely specify the a folder with f to pre-load, and the default tfs are always present. 

Add two optional parameters to the configs allowing loading pre-registered tf through the path attribute of STransferFunction.

Add support to empty path (default value in the config) in the STransferFunction service.

Changes the TransferFunctionWidgetCfg and TransferFunctionWindowsCfg config id to fetch new standards (sight::config::viz::scene2d::TransferFunctionWidgetCfg)

Closes #1065