enh(viz): add reset layer camera slot
The reset camera slot is connected, on the SRender service, to all layouts of the service. 
However, when several layouts are present, it can be expected to reset only a given layout's camera. 

To do this, a dedicated slot is created at the service configuration. 
The slot name contains the layoutId, and call the resetCameraCoordinates which is specified for the layoutId provided in parameter. 

Closes #1066