feat(ui): provide a view that can be overlayed on 3D scenes
A new `OverlayLayoutManager` is introduced, allowing the addition of overlays to a widget. Here is an example of use:

```xml
<service uid="mainView" type="sight::module::ui::base::SView">
    <gui>
        <layout type="sight::ui::base::OverlayLayoutManager">
            <view />
            <view x="0" y="0" minWidth="55" minHeight="100" />
            <view x="-1" y="0" width="50" height="55" visible="false" />
            <view x="0" y="-1" height="35" width="100%" visible="false" />
            <view x="-50%" y="0" width="400" height="300" />
        </layout>
    </gui>
    <registry>
        <parent wid="${WID_PARENT}" />
        <view sid="scenesView" start="true" />
        <view sid="topToolbarSliderView" start="true" />
        <view sid="rightToolbarSliderView" start="true" />
        <view sid="videoSliderView" start="true" />
        <view sid="advancedQueryEditorSliderView" start="true" />
    </registry>
</service>
```
In OverlayLayoutManager, we define a list of views. The first view is the background widget, typically a 3D scene. The other ones are overlays, we can define properties such as ``x``, ``y``, ``width``, ``height`` to indicate where the overlay should be drawn and what its size should be. By default, ``x`` and ``y`` are the offset from the left and the top respectively, however, if they are negative, then it is the offset from the right and the bottom respectively. By default, they are absolute sizes in pixels, however, if ``%`` is appended, then it is relative to the size of the parent widget.