fix(io): add openJpg support for dicom writer
Some dicom tests writing fail because of an internal mismatch in dicom/nvJpeg2000/openJpeg.

To summarize, for strongly randomized images, nvpeg2000 fails. This has been fixed in the 0.7 version of nvjpeg, used on linux, but not presently on windows.
So, when writing the dicom in a ut on windows, nvJpeg2000 is detected, used, fails, and a backup is used to openJpeg.

However, gdcm, the dicom writing lib, uses openJpeg intrinsecly, and does not support an the external use of openjpeg (bug ! ) which lead,s in the present case to a fail of the nvjpeg2000 backup system.

As a result, some tests fail on windows.

To fix this, a support of dicom writing in onpenjpeg has been added, to make syure that the dicom is writen in any cases