feat(ui): create buttonbar widget for SParam enum
The SParameter supports enums, though they are not simple to use, in terms of xml and ui.
A new  enum display will be added, corresponding to a bar, with buttons, image and text which allow to select a specific value.

It should have the features of enumSParam, with icons and labels. 
The unselected options are gray, while the selected one is in full color. 

Gui test are added, to check this feature, and a sample is added in SParameter.