fix(viz): shape extruder does not work with touchscreen
When using shape extruder on a volume in a 3D scene, and using a mousse, all works fine. 
However, on the same machine, same volume, same extrusion path, when doing it with the touchscreen the extrusion takes ages, and freezes the application. 

This has been fixed by the way events are handled. 

shape extrusion should work perfectly on both touchscreen and mousse actions.