enh(viz): improve distance measurement tool with a click and drag movement
The distance measurement tool had some limitations. 
In particular, at the creation, the initial process was : 

- click on add distance: a distance is created at the ends (0,0) and (width, height) of the image
- drag-drop the created distance ends to the location you want to measure

This was too tedious. The new behaviour is : 

- click on add Distance: enter a dedicated mode
- click on the starting point you want to measure
- drag to the end
- drop, the distance is ready. 

During your measurement, the distance is automatically updated on the screen, allowing you to visualise what you measure, and the distance between the starting point and present location. 

SightViewer has been updated to support the new tool and to have the negato2D and 3D interacting when the distance is created and measured.