fix(viz): make SCamera calibrate safer
The SCamera adaptor requires a calibration step.
This calibration can be provided through an input to the service (CameraSet, cameraCalibration). However, for virtual camera in classical 3Dscene, no calibration is provided, and the calibration is computed based on the viewport size.
This set is only called at the service start, and, if the viewport is hidden, or deffined with visibility=false, its height and with are null or negativ.
This leads to an error which stops the application in debug.

Add a calibration verification checK.
If the calibration fails in SCamera, it stops the updating.
If the updating is called and the calibration has not been done, it is calibrated again.

Closes #1135