enh(ui): rework view maximization
This fixes the standard maximization shortcut (F11 on Windows, on Linux it depends on the window manager) when a 3d scene is displayed. The 3D scene used its own QShortcut that took precedence over the one installed on the main frame.

This also updates sight_viewer to use a more common maximization/restore behavior, but since the 3D scene real full-screen has also been removed, the tutorial "07" has also been changed to include a demonstration of the 3D scene full-screen.

And last but not the least, a crash was fixed in Speeddial code, discovered while modifying `tuto07_generic_scene`. Another kind of "deadlock" (window that was never closed) was also fixed in gui tests (seen in sight_viewer_uit)