refactor(core): do not enable all modules by default
Sight used to provide a mechanism to select the modules for an application. This is the `<activate>` tag of the `profile.xml`. A few years ago, it was decided, by simplicity, to enable all modules by default. Indeed, when an app is installed, the selection is somehow done thanks to the CMake dependencies. However, in the build tree, you may have access to modules your app does not depend on. This is not that tragic, but sometimes this is annoying because we realize we have a missing dependency against an XML configuration for instance only at install time. Also, we may have more extensions available (for instance readers) in the build tree than in the install tree.

With this merge-request, only the modules that an app depends on (recursively) are enabled.