fix(core): setup the correct profile version for each app
The version of the application is now well generated in the `profile.xml` of each app. Thus, it is now safe to use:

```cpp
const auto profile     = sight::core::runtime::get_current_profile();
const software_version = profile->version();
```

It is even recommended to use this instead of passing the APP_VERSION through the XML or by a C++ compiler macro.