enh(core): introduce explicit update loops
Chaining service updates in Sight live applications has always been tough. To circumvent all related issues with event-based chaining, we introduce an explicit approach allowing to specify update sequence and execute them all at once.

It used to be possible to specify which services to update **once** once after start:

```xml
<update uid="service1">
<update uid="service2">
<update uid="service3">
```

This section is now modified and allows specifying these single-shot updates, but also the ability to define an update loop:

```xml
<start uid="update_loop">

<update>
     <service uid="service1"/>
     <service uid="service2"/>
     <sequence uid="update_loop" loop="true">
         <service uid="service3" />
         <service uid="service4" />
     </sequence>
</update>
```

In this example, `service1` and `service2` are updated once after start, while `service3` and `service4` are continuously updated. The sequence `update_loop`is identified with an uid, so it can be started and stopped like any other service. Doing this, the update sequence is not dependent on the updated() signal. This is much more robust because this means any other, maybe unwanted updates triggered from somewhere else do not affect the sequence.

This new syntax also allows chaining update sequences between XML configurations. As the manner of UI registries and `wid`, it is possible to reserve a slot for another update sequence defined in another configuration:

```xml
<!-- Main loop -->
<update>
     <sequence uid="update_loop" loop="true">
         <service uid="service1" />
         <updater uid="sub_updater" />
         <service uid="service2" slot="modify" />
     </sequence>
</update>

<!-- In a sub-configuration, passing the parameter "sub_updater" -->
<update>
     <sequence parent="${sub_updater}">
         <service uid="..." />
         <service uid="..." />
     </sequence>
</update>
```

Note that here the `uid` of the sequence is omitted because no interaction is required inside this XML. The sequence is played automatically by the root updater after `service1::update()` and `service2::modify` slots.

It is also possible to execute services in parallel. Mixing sequences and parallel are allowed with the syntax:

```xml
<!-- Main loop -->
<update>
     <sequence uid="update_loop" loop="true">
         <service uid="service1" />
         <parallel>
            <service uid="service2" />
            <updater uid="sub_updater" />
            <sequence>
               <service uid="service3" />
               <service uid="service4" />
            </sequence>
         <parallel/>
         <service uid="service5" />
     </sequence>
</update>

<!-- In a sub-configuration, passing the parameter "sub_updater" -->
<update>
     <sequence parent="${sub_updater}">
         <service uid="..." />
         <service uid="..." />
     </sequence>
</update>
```

In this example, `service1`updates first, then `service2`, the whole update sequence in the sub-configuration, and the sequence of `service3` and `service4` are executed in parallel. When they are executed, whether or not they are on different workers, `service5` is executed.

The solution proposed above could be sufficient. However, managing `sight::viz::scene3d::render` could become tedious. Correctly using auto-connections, and maintaining the list of all adaptors in the update can be tedious and error-prone. Developers usually expect that a single call to the \`update()\` of the renderer manages the update of the adaptors properly, which is the case in auto mode when we do interactive rendering, but not in manual mode, in "live" applications.

To solve this, we introduced a generic update interface `sight::core::updater` that will be inherited by the adaptors. This interface allows to defer the update of the adaptor required by their data. In manual mode, the adaptors just flag which part of their update process they should process. When the update is actually called, they will really perform it and unflag it. The render only requires to interrogate the adaptors before rendering to know if they need to update or not.

To sum up, updating a generic scene in manual mode is now much easier and not error-prone regarding synchronisation. For example, if we imagine an application reading a video, extracting a feature and then display it in overlay over the video, the update loop could look like:

```xml
<!-- Main loop -->
<update>
     <sequence uid="main_loop" loop="true">
         <service uid="video_grabber_srv" />
         <service uid="extract_position_srv" />
         <service uid="compute_position_srv" />
         <service uid="render_srv" /> <!-- Update all adaptors, only those required if their data have changed -->
     </sequence>
</update>
```