fix(core): improve handling of connection exceptions in proxies
Now, when an incompatibility is found between the signal and the slot, the exception is caught in the proxy and an error is returned. The config can still run, but it will not crash when closing.

Also, an important bug was found when trying to match signal and slot signature. We can indeed omit part of the arguments of a signal when connecting a slot, but of course from the right. We can connect for instance:
- (int, int) -> (int)
- (int, int) -> ()

But of course we refuse the inverse (we throw `sight::core::com::exception::bad_slot`):
- (int) -> (int, int)
- () -> (int, int)

This mechanism works recursively but was broken when argument types were different like:
- (int, int) -> (string)

This led to a stack overflow. :confused: This is now fixed and throw `sight::core::com::exception::bad_slot` as well.