feat(core): add a new service to reset any data
So far, we have had no service to reset any data object to the default. We used a `sight::data::module::copy` service with a default "empty" object to achieve that.

The new service `sight::data::module::reset` does this in a simpler way, because empty objects in XML are thus no longer necessary.

```xml
<service uid="..." type="sight::module::data::reset" >
    <inout key="target" uid="..." />
</service>
```