fix(core): various sequencer and activity corrections
First, a new object validator that verifies if an object is empty. For now it handles `sight::data::image` and `sight::data::point_list`. It is likely to be extended to support more object types in the future. 
Then some cleaning was required in the configuration manager to make recent features compatible with activities. For instance, we require activity objects to have a real unique identifier that matches the requirement names. This is mandatory to get those objects work with the recent `sight::ui::module_qt::settings` service.