feat(core): make object validator generic, add service to validate
A new service `sight::module::data::validate` was implemented to validate objects in a generic way. Here is an example to check that an image is filled:

```xml
<service uid="..." type="sight::module::data::validate" >
    <in key="data" uid="image"/>
    <config id="sight::data::validator::filled" />
</service>
```

It provides the following signals in XML, which can be connected to your application logic:

* valid()
* invalid()
* is_valid(bool)
* is_invalid(bool)

Also, a new validator `sight::data::validator::has_fiducials` was implemented to check if a series contains at least one fiducial.