refactor(core): autostart services in XML and simplify config launcher
Services are now autostarted in XML. Since most services required to be started, it is simpler to start them automatically.

For the rare services that do not start, you can use a new property `start`:

```xml
<service uid="..." type="sight::module::ui::action">
    <properties start="false" />
</service>
```

You can then start/stop the service via slots, as usual. But of course, since this is a property, you can also use a variable:

```xml
<service uid="..." type="sight::module::ui::action">
    <properties start="${start_action}" />
</service>
```

Other changes:
* Renamed the `sight::app::config_controller` into `sight::app::config_launcher`
* Replaced all usages of `sight::module::ui::config_launcher` with `sight::app::config_launcher`
* Removed `sight::module::ui::config_launcher` service